requirements.txt
Django==1.10.5
httplib2==0.10.3
olefile==0.44
Pillow==4.1.0
python-http-client==2.2.1
python-instagram==1.3.2
requests==2.13.0
sendgrid==4.0.0
simplejson==3.10.0
six==1.10.0
